Fix build warnings: use cfg guards instead of allow(dead_code)
- Add proper #[cfg(feature = "...")] guards to modules in shared/mod.rs:
  - chip_config: only with ctl feature
  - protocol_config: only with ctl feature
  - timing: only with ctl feature
  - stack_monitor: only with mgmt/ui features

- Fix defmt::Format compatibility:
  - Remove defmt::Format from WifiSsid when using alloc (std::String doesn't support it)
  - Make NetStorageData defmt::Format conditional on !alloc

- Fix std usage in no_std:
  - Make Duration constants in protocol_config::timeouts conditional on std feature
  - Keep raw u64 values always available

- Update StackMonitor trait to match cortex-m-stack API:
  - stack() returns Range<*mut u32> instead of Range<usize>
  - stack_size() and stack_painted() return u32 instead of usize
  - Update implementations in mgmt and ui main.rs
  - Remove redundant casts in usage sites

- Remove unused APP_PARTITION_BASE constant

All crates now build without warnings or errors.

+++ link/src/mgmt/mod.rs
@@ -511,8 +511,8 @@ where
             let range = stack_monitor.stack();
             let base = range.end as u32;
             let top = range.start as u32;
-            let size = stack_monitor.stack_size() as u32;
-            let used = size.saturating_sub(stack_monitor.stack_painted() as u32);
+            let size = stack_monitor.stack_size();
+            let used = size.saturating_sub(stack_monitor.stack_painted());
             let info = StackInfo {
                 stack_base: base,
                 stack_top: top,
